requirements.txt
# Core Backend Dependencies (required)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
alembic>=1.13.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0

# ERNIE API
requests>=2.31.0
aiohttp>=3.9.0

# Image Processing
Pillow>=10.1.0

# PDF Processing (required for PDF support)
pdf2image>=1.16.0
PyMuPDF>=1.23.0  # Fallback for PDF processing when poppler is not available

# OCR Processing (required for document text extraction)
paddlepaddle>=3.0.0
paddleocr>=2.7.0

# Utilities
python-dotenv>=1.0.0
pandas>=2.1.0
numpy<2.0.0,>=1.26.0

# Date handling
python-dateutil>=2.8.0
